+++ klibs/tests/test_KLBoundary.py
@@ -13,11 +13,16 @@ def test_rectangle_boundary():
     pos = klb.RectangleBoundary('test2', (10, 10), (50, 50))
     floats = klb.RectangleBoundary('test3', p1=(10.4, 10.8), p2=(50.5, 50.2))
 
+    # Test string
+    assert str(rect) == "RectangleBoundary(p1=(10, 10), p2=(50, 50))"
+
     # Test boundary attributes
     assert rect.label == 'test1'
     assert rect.p1 == (10, 10)
     assert rect.p2 == (50, 50)
     assert rect.center == (30, 30)
+    assert rect.width == 40
+    assert rect.height == 40
 
     # Test boundary usage
     assert rect.within((0, 0)) == False
@@ -25,12 +30,31 @@ def test_rectangle_boundary():
     assert rect.within((35.3, 35)) == True
     assert rect.within((10, 10)) == True
     assert rect.within((50, 50)) == True
+    assert (10, 10) in rect
+    assert not (0, 0) in rect
+
+    # Test boundary movement
+    rect.center = (50, 60)
+    assert rect.p1 == (30, 40)
+    assert rect.p2 == (70, 80)
+    assert rect.within((10, 10)) == False
+    assert rect.within((70, 80)) == True
+
+    # Test handling of when p1 > p2
+    swapped = klb.RectangleBoundary('swapped', (50, 50), (10, 10))
+    assert swapped.p1 == (10, 10)
+    assert swapped.p2 == (50, 50)
+    swapped = klb.RectangleBoundary('swapped', (50, 10), (10, 50))
+    assert swapped.p1 == (10, 10)
+    assert swapped.p2 == (50, 50)
 
     # Test boundary exceptions
     with pytest.raises(ValueError):
         rect.within(5)
     with pytest.raises(ValueError):
-        klb.RectangleBoundary('test4', p1=(60, 60), p2=(50, 50))
+        rect.center = (0, 0, 0)
+    with pytest.raises(ValueError):
+        klb.RectangleBoundary('test4', p1=(60, 60), p2=(60, 50))
     with pytest.raises(ValueError):
         klb.RectangleBoundary('test5', p1=0, p2=(50, 50))
 
@@ -42,6 +66,9 @@ def test_circle_boundary():
     # Test position arguments and boundaries with floats
     pos = klb.CircleBoundary('test2', (100, 100), 50)
     floats = klb.CircleBoundary('test3', center=(99.5, 100), radius=43.5)
+
+    # Test string
+    assert str(circle) == "CircleBoundary(center=(100, 100), radius=50)"
 
     # Test boundary attributes
     assert circle.label == 'test1'
@@ -55,10 +82,20 @@ def test_circle_boundary():
     assert circle.within((99.7, 100)) == True
     assert circle.within((50, 100)) == True
     assert circle.within((100, 150)) == True
+    assert (50, 100) in circle
+    assert not (0, 0) in circle
+
+    # Test boundary movement
+    circle.center = (50, 50)
+    assert circle.center == (50, 50)
+    assert circle.within((100, 150)) == False
+    assert circle.within((50, 1)) == True
 
     # Test boundary exceptions
     with pytest.raises(ValueError):
         circle.within(5)
+    with pytest.raises(ValueError):
+        circle.center = (0, 0, 0)
     with pytest.raises(ValueError):
         klb.CircleBoundary('test4', center=(100, 100), radius=-4)
     with pytest.raises(ValueError):
@@ -72,6 +109,9 @@ def test_annulus_boundary():
     # Test position arguments and boundaries with floats
     pos = klb.AnnulusBoundary('test2', (100, 100), 50, 10)
     floats = klb.AnnulusBoundary('test3', center=(99.5, 100), radius=43.5, thickness=4.6)
+
+    # Test string
+    assert str(ring) == "AnnulusBoundary(center=(100, 100), radius=50, thickness=10)"
 
     # Test boundary attributes
     assert ring.label == 'test1'
@@ -91,10 +131,20 @@ def test_annulus_boundary():
     assert ring.within((60, 100)) == True
     assert ring.within((100, 150)) == True
     assert ring.within((100, 140)) == True
+    assert (55, 100) in ring
+    assert not (0, 0) in ring
+
+    # Test boundary movement
+    ring.center = (50, 50)
+    assert ring.center == (50, 50)
+    assert ring.within((100, 150)) == False
+    assert ring.within((50, 1)) == True
 
     # Test boundary exceptions
     with pytest.raises(ValueError):
         ring.within(5)
+    with pytest.raises(ValueError):
+        ring.center = (0, 0, 0)
     with pytest.raises(ValueError):
         klb.AnnulusBoundary('test4', center=(100, 100), radius=-4, thickness=10)
     with pytest.raises(ValueError):
@@ -105,9 +155,9 @@ def test_annulus_boundary():
         klb.AnnulusBoundary('test7', center=100, radius=50, thickness=10)
 
 
-def test_boundary_inspector():
+def test_boundary_set():
 
-    inspector = klb.BoundaryInspector()
+    inspector = klb.BoundarySet()
     tst1 = klb.RectangleBoundary('test1', p1=(10, 10), p2=(50, 50))
     tst2 = klb.RectangleBoundary('test2', p1=(60, 60), p2=(100, 100))
     tst3 = klb.RectangleBoundary('test3', p1=(10, 30), p2=(50, 70))
@@ -126,15 +176,14 @@ def test_boundary_inspector():
     assert len(inspector.boundaries) == 3
 
     # Test current method of adding boundaries to inspector
-    inspector = klb.BoundaryInspector()
+    inspector = klb.BoundarySet()
     inspector.add_boundary(tst1)
     assert len(inspector.boundaries) == 1
     inspector.add_boundaries([tst2, tst3])
     assert len(inspector.boundaries) == 3
 
     # Test removing boundaries from the inspector
-    inspector = klb.BoundaryInspector()
-    inspector.add_boundaries([tst1, tst2, tst3])
+    inspector = klb.BoundarySet([tst1, tst2, tst3])
     assert len(inspector.boundaries) == 3
     inspector.remove_boundaries('test1')
     assert len(inspector.boundaries) == 2
@@ -142,24 +191,39 @@ def test_boundary_inspector():
     assert len(inspector.boundaries) == 0
 
     # Test clearing boundaries from the inspector
-    inspector = klb.BoundaryInspector()
+    inspector = klb.BoundarySet()
     inspector.add_boundaries([tst1, tst2, tst3])
     inspector.clear_boundaries()
     assert len(inspector.boundaries) == 0
     inspector.add_boundaries([tst1, tst2, tst3])
     inspector.clear_boundaries(preserve=['test2'])
     assert len(inspector.boundaries) == 1
-    assert 'test2' in inspector.boundaries.keys()
+    assert 'test2' in inspector.labels
+    
+    # Test that preserved boundaries during clear keep same order
+    inspector.add_boundaries([tst1, tst2, tst3])
+    inspector.clear_boundaries(preserve=['test3', 'test1'])
+    assert len(inspector.boundaries) == 2
+    assert inspector.labels == ['test1', 'test3']
 
     # Test individual boundary tests
-    inspector = klb.BoundaryInspector()
+    inspector = klb.BoundarySet()
     inspector.add_boundaries([tst1, tst2, tst3])
     assert inspector.within_boundary('test1', (20, 40)) == True
     assert inspector.within_boundary('test2', (20, 40)) == False
     assert inspector.within_boundary('test3', (20, 40)) == True
 
+    # Test combined boundary tests
+    inspector = klb.BoundarySet()
+    inspector.add_boundaries([tst1, tst2, tst3])
+    assert inspector.which_boundary((20, 40)) == 'test3'
+    assert inspector.which_boundary((20, 40), ignore='test3') == 'test1'
+    assert inspector.which_boundary((20, 40), ignore=['test3']) == 'test1'
+    assert inspector.which_boundary((20, 40), labels=['test1', 'test2']) == 'test1'
+    assert inspector.which_boundary((20, 40), labels=['test2']) == None
+
     # Test exceptions
-    inspector = klb.BoundaryInspector()
+    inspector = klb.BoundarySet()
     with pytest.raises(ValueError):
         inspector.add_boundary('hello')
     with pytest.raises(ValueError):